dbrepot t5 g

+++ src/test/java/app/domain/wiseSaying/repository/WiseSyaingDbRepositoryTest.java
@@ -117,7 +117,7 @@ public class WiseSyaingDbRepositoryTest {
         wiseSayingDbRepository.save(wiseSaying2);
 
 
-        int count = wiseSayingDbRepository.count();
+        long count = wiseSayingDbRepository.count();
 
         assertThat(count)
                 .isEqualTo(2);
